User Registry Migrations code updated, models seperated.
