+++ app/__init__.py
@@ -10,4 +10,7 @@ app = Flask(__name__)
 app.config.from_object(os.getenv('APP_SETTINGS'))
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
+
+from app.models import user, blackList, domainCheck
+
 migrate = Migrate(app, db)
